Moved test classes to maven test folder. Fixed bug in test
