+++ src/test/java/org/tuberlin/de/common/ModelTest.java
@@ -1,4 +1,4 @@
-package org.tuberlin.de.common.test;
+package org.tuberlin.de.common;
 
 /**
  * Created by Malcolm-X on 26.12.2015.
